requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
# DB drivers
psycopg2-binary==2.9.9
pymssql==2.3.0
pymysql==1.1.0
pymongo==4.7.2
# SQLite built-in, no package needed
# Encryption + security
cryptography==42.0.8
# Persistence
sqlalchemy==2.0.30
# Similarity
rapidfuzz==3.9.3
# Utilities
python-dotenv==1.0.1
# Graph
neo4j==5.20.0
